Conflicts:
	arangod/VocBase/collection.c
	arangod/VocBase/compactor.c
	arangod/VocBase/primary-collection.c
	arangod/VocBase/shape-collection.c

+++ arangod/VocBase/primary-collection.c
@@ -128,7 +128,6 @@ static TRI_datafile_t* CreateJournal (TRI_primary_collection_t* primary, bool co
 
     // construct a suitable filename
     number = TRI_StringUInt64((uint64_t) TRI_NewGlobalIdSequence());
-
     if (compactor) {
       jname = TRI_Concatenate3String("journal-", number, ".db");
     }
@@ -171,8 +170,7 @@ static TRI_datafile_t* CreateJournal (TRI_primary_collection_t* primary, bool co
     bool ok;
 
     // and use the correct name
-    number = TRI_StringUInt64((uint64_t) journal->_fid);
-
+    number = TRI_StringUInt64(journal->_fid);
     if (compactor) {
       jname = TRI_Concatenate3String("compactor-", number, ".db");
     }
@@ -291,7 +289,7 @@ static bool CloseJournalPrimaryCollection (TRI_primary_collection_t* primary,
     char* number;
     bool ok;
 
-    number = TRI_StringUInt64((uint64_t) journal->_fid);
+    number = TRI_StringUInt64(journal->_fid);
     dname = TRI_Concatenate3String("datafile-", number, ".db");
     filename = TRI_Concatenate2File(collection->_directory, dname);
